separacion de funciones de uso comun en otro archivo Commonfunctions
